multiple row solve
Here I have made it so that the system will solve for each row in the square using the first algorithm I have implemented. This required a few adjustments but the most notable change was getting rid of the cells array in the value class as I dont need it and cant adjust anything into it anyways. this might change though